Front/fix/modify ui

See merge request s04-webmobile2-sub3/s04p13c108!313

+++ frontend/src/components/group/FeedCard.vue
@@ -154,12 +154,15 @@ import { notiType, sendNotify } from "../../api/notification/index"
         
       },
       feedDelete(){
+      if(confirm("피드를 삭제하시겠습니까?")==true) {
         http.delete(`feed/deleteFeed?feedPk=${this.card.feedPk}`)
         .then(()=>{
+          alert("피드가 삭제되었습니다.")
           this.$router.go()
-          alert("삭제되었습니다.")
           }
         )
+      }         
+        
       },
       feedUpdate(){
         this.$router.push({name:"FeedUpdate",params:{feedPk:Number(this.card.feedPk),"content":this.card.content, "writer":this.card.writer, "image":this.card.media}})
